+++ src/app/state/pagination.rs
@@ -1,11 +1,11 @@
 pub struct Pagination<T> {
     pub data: T,
-    pub next_offset: Option<u32>,
-    pub batch_size: u32,
+    pub next_offset: Option<usize>,
+    pub batch_size: usize,
 }
 
 impl<T> Pagination<T> {
-    pub fn new(data: T, batch_size: u32) -> Self {
+    pub fn new(data: T, batch_size: usize) -> Self {
         Self {
             data,
             next_offset: Some(0),
@@ -13,7 +13,7 @@ impl<T> Pagination<T> {
         }
     }
 
-    pub fn reset_count(&mut self, new_length: u32) {
+    pub fn reset_count(&mut self, new_length: usize) {
         self.next_offset = if new_length >= self.batch_size {
             Some(self.batch_size)
         } else {
@@ -21,7 +21,7 @@ impl<T> Pagination<T> {
         }
     }
 
-    pub fn set_loaded_count(&mut self, loaded_count: u32) {
+    pub fn set_loaded_count(&mut self, loaded_count: usize) {
         if let Some(offset) = self.next_offset.take() {
             self.next_offset = if loaded_count >= self.batch_size {
                 Some(offset + self.batch_size)
